TodoCounter added and checkbox logic working.
Created a TodoCount component that is basically just responsible for display purposes. The TodoList displays the list from the api and the checkbox next to each Todo title. When the checkbox is checked, the handleCheckboxChange function checks the document for all instances of a checked checkbox and then loads that number to the checkedTodos useState.