updated handleCheckboxChange logic to use useRef instead querySelectorAll().
Completely reworked the handleCheckboxChange logic. Now it uses a useRef function. When a checkbox is checked useRef initializes and creates a place for the matching todo.id. It is then set to checked. This is then converted to a number and useState is updated for checkedTodos that will then feed the number to the TodoCounter.